+++ src/vm/errors/vm_errors.rs
@@ -31,11 +31,11 @@ pub enum VirtualMachineError {
     UnconstrainedResJumpRel,
     #[error("Res.UNCONSTRAINED cannot be used with Opcode.ASSERT_EQ")]
     UnconstrainedResAssertEq,
-    #[error("ASSERT_EQ instruction failed; {0:?} != {1:?}")]
+    #[error("An ASSERT_EQ instruction failed: {0} != {1}.")]
     DiffAssertValues(MaybeRelocatable, MaybeRelocatable),
-    #[error("Call failed to write return-pc (inconsistent op0): {0:?} != {1:?}. Did you forget to increment ap?")]
+    #[error("Call failed to write return-pc (inconsistent op0): {0} != {1}. Did you forget to increment ap?")]
     CantWriteReturnPc(MaybeRelocatable, MaybeRelocatable),
-    #[error("Call failed to write return-fp (inconsistent dst): {0:?} != {1:?}. Did you forget to increment ap?")]
+    #[error("Call failed to write return-fp (inconsistent dst): {0} != {1}. Did you forget to increment ap?")]
     CantWriteReturnFp(MaybeRelocatable, MaybeRelocatable),
     #[error("Couldn't get or load dst")]
     NoDst,
@@ -53,11 +53,11 @@ pub enum VirtualMachineError {
     NotImplemented,
     #[error("Can only subtract two relocatable values of the same segment")]
     DiffIndexSub,
-    #[error("Inconsistent auto-deduction for builtin {0}, expected {1:?}, got {2:?}")]
+    #[error("Inconsistent auto-deduction for builtin {0}, expected {1}, got {2:?}")]
     InconsistentAutoDeduction(String, MaybeRelocatable, Option<MaybeRelocatable>),
     #[error(transparent)]
     RunnerError(#[from] RunnerError),
-    #[error("Invalid hint encoding at pc: {0:?}")]
+    #[error("Invalid hint encoding at pc: {0}")]
     InvalidHintEncoding(MaybeRelocatable),
     #[error(transparent)]
     MemoryError(#[from] MemoryError),
@@ -65,11 +65,11 @@ pub enum VirtualMachineError {
     NoRangeCheckBuiltin,
     #[error("Expected ecdsa builtin to be present")]
     NoSignatureBuiltin,
-    #[error("Failed to retrieve value from address {0:?}")]
+    #[error("Failed to retrieve value from address {0}")]
     MemoryGet(MaybeRelocatable),
-    #[error("Expected integer at address {0:?}")]
+    #[error("Expected integer at address {0}")]
     ExpectedInteger(MaybeRelocatable),
-    #[error("Expected relocatable at address {0:?}")]
+    #[error("Expected relocatable at address {0}")]
     ExpectedRelocatable(MaybeRelocatable),
     #[error("Failed to get ids for hint execution")]
     FailedToGetIds,
@@ -77,7 +77,7 @@ pub enum VirtualMachineError {
     NonLeFelt(Felt, Felt),
     #[error("Div out of range: 0 < {0} <= {1}")]
     OutOfValidRange(Felt, Felt),
-    #[error("Assertion failed, 0 <= ids.a % PRIME < range_check_builtin.bound \n a = {0:?} is out of range")]
+    #[error("Assertion failed, 0 <= ids.a % PRIME < range_check_builtin.bound \n a = {0} is out of range")]
     ValueOutOfRange(Felt),
     #[error("Value: {0} should be positive")]
     ValueNotPositive(Felt),
@@ -89,11 +89,11 @@ pub enum VirtualMachineError {
     SplitIntNotZero,
     #[error("split_int(): Limb {0} is out of range.")]
     SplitIntLimbOutOfRange(Felt),
-    #[error("Failed to compare {0:?} and {1:?}, cant compare a relocatable to an integer value")]
+    #[error("Failed to compare {0} and {1}, cant compare a relocatable to an integer value")]
     DiffTypeComparison(MaybeRelocatable, MaybeRelocatable),
-    #[error("assert_not_equal failed: {0:?} =  {1:?}")]
+    #[error("assert_not_equal failed: {0} =  {1}")]
     AssertNotEqualFail(MaybeRelocatable, MaybeRelocatable),
-    #[error("Failed to compare {0:?} and  {1:?}, cant compare two relocatable values of different segment indexes")]
+    #[error("Failed to compare {0} and  {1}, cant compare two relocatable values of different segment indexes")]
     DiffIndexComp(Relocatable, Relocatable),
     #[error("Value: {0} is outside of the range [0, 2**250)")]
     ValueOutside250BitRange(Felt),
@@ -137,7 +137,7 @@ pub enum VirtualMachineError {
     NoneApTrackingData,
     #[error("Tracking groups should be the same, got {0} and {1}")]
     InvalidTrackingGroup(usize, usize),
-    #[error("Expected relocatable for ap, got {0:?}")]
+    #[error("Expected relocatable for ap, got {0}")]
     InvalidApValue(MaybeRelocatable),
     #[error("Dict Error: Tried to create a dict whithout an initial dict")]
     NoInitialDict,
@@ -183,13 +183,13 @@ pub enum VirtualMachineError {
     CouldntPopPositions,
     #[error("unexpected verify multiplicity fail: last_pos not found")]
     LastPosNotFound,
-    #[error("Set's starting point {0:?} is bigger it's ending point {1:?}")]
+    #[error("Set's starting point {0} is bigger it's ending point {1}")]
     InvalidSetRange(MaybeRelocatable, MaybeRelocatable),
     #[error("Failed to construct a fixed size array of size: {0}")]
     FixedSizeArrayFail(usize),
     #[error("{0}")]
     AssertionFailed(String),
-    #[error("Wrong dict pointer supplied. Got {0:?}, expected {1:?}.")]
+    #[error("Wrong dict pointer supplied. Got {0}, expected {1}.")]
     MismatchedDictPtr(Relocatable, Relocatable),
     #[error("Integer must be postive or zero, got: {0}")]
     SecpSplitNegative(BigInt),
@@ -223,7 +223,7 @@ pub enum VirtualMachineError {
     NoImm,
     #[error("Tried to compute an address but there was no register in the reference.")]
     NoRegisterInReference,
-    #[error("Couldn't compute operand {0} at address {1:?}")]
+    #[error("Couldn't compute operand {0} at address {1}")]
     FailedToComputeOperands(String, Relocatable),
     #[error("Custom Hint Error: {0}")]
     CustomHint(String),