+++ CHANGELOG.md
@@ -5,3 +5,8 @@
         * `pub fn get_location(pc: &usize, runner: &CairoRunner) -> Option<Location>` is now `pub fn get_location(pc: usize, runner: &CairoRunner) -> Option<Location>`
         * `pub fn decode_instruction(encoded_instr: i64, mut imm: Option<BigInt>) -> Result<instruction::Instruction, VirtualMachineError>` is now `pub fn decode_instruction(encoded_instr: i64, mut imm: Option<&BigInt>) -> Result<instruction::Instruction, VirtualMachineError>`
         * `VmExcepion` field's string format now mirror their cairo-lang conterparts.
+
+* Add input file contents to traceback [#666](https://github.com/lambdaclass/cairo-rs/pull/666/files)
+    * Public Api changes:
+        * `VirtualMachineError` enum variants containing `MaybeRelocatable` and/or `Relocatable` values now use the `Display` format instead of `Debug` in their `Display` implementation
+        * `get_traceback` now adds the source code line to each traceback entry